+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * **Breaking change**: New external signing mode without the need for creating
   the PKCS#7/CMS signed data object for
   [HexaPDF::DigitalSignature::Signing::DefaultHandler]
+* **Breaking change**: Use value :pades instead of :etsi for
+  [HexaPDF::DigitalSignature::Signing::DefaultHandler#signature_type]
 * [HexaPDF::DigitalSignature::Signing::DefaultHandler] to allow specifying the
   used digest algorithm
 * [HexaPDF::DigitalSignature::Signing::DefaultHandler] to allow specifying a